Listen on both IPv4 and IPv6.
Recent changes to Docker have made it so exposed ports that are bound to IPv6 will only forward to the IPv6 container address. This now requires the code in the container to listen on IPv6.

The "v4v6" part ensures that HAProxy will also listen on IPv4, regardless of system/container config.